feat(social-login): Enable social login
Currently, the only way a user gains access to the system is
if they register/login using the login and register form
which is inconvinient since the user has to remember their
credentials.

This commit integrates the ability to login using social
media accounts ie. twitter, facebook and google which
improves the UX.

[Delivers #157731629]