feat(Get-exercises-on-muscle-hover) Creates a get exercises on hover over body part feature
This commit creates a feature that enables users to hover over a body part in the image
and display the exercises for that body part.

Add event handlers to wger/exercises/templates/muscles/overview.html
to get the required target on mouse hover over image, and display the relevant body exercises
Add flip-front and flip-back button to enable users flip the image to the back side,
in case you want to hover over the back muscles
Add default svg css to wger/core/static/css/workout-manager.css to hide the overlapping svg
images with the targeted ids and position the image over the background image.

[Delivers #157731630]