feat(relation btn log and session): Add relation between log and session
Originally, workout logs and the workout session are only linked
through the date (for sessions this value is unique and enforced
by the DB). This presents in some places a problem, since it's
not possible to just follow a foreign key relationship with the
ORM like in other places.

This commit creates a log relation between the WorkoutLog and
WorkoutSession

[Delivers #157731617]